\name{JGL2-package}
\alias{JGL2-package}
\alias{JGL2}
\docType{package}
\title{
\packageTitle{JGL2}
}
\description{
\packageDescription{JGL2}
}
\author{
\packageAuthor{JGL2}
}
\references{
Danaher, P., Wang, P., and Witten, D. M. (2014). The joint graphical lasso for inverse covariance estimation across multiple classes. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 76(2), 373-397. http://doi.org/10.1111/rssb.12033
Guo, J., Levina, E., Michailidis, G., and Zhu, J. (2011). Joint estimation of multiple graphical models. Biometrika, 98(1), 1-15. http://doi.org/10.1093/biomet/asq060
}

\seealso{
\code{\link{JGL}}
}


